+++ examples/guides/link_provider_to_ou.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     clumio = {
-      source  = "clumio.com/providers/clumio"
+      source  = "clumio-code/clumio"
       version = "~>0.4.0"
     }
   }